Replace old modal instances with new modal
