+++ web-common/src/features/welcome/TitleContent.svelte
@@ -49,11 +49,10 @@
       Add data
     </div>
   </button>
-  {#if showAddSourceModal}
-    <AddSourceModal
-      on:close={() => {
-        showAddSourceModal = false;
-      }}
-    />
-  {/if}
+  <AddSourceModal
+    open={showAddSourceModal}
+    on:close={() => {
+      showAddSourceModal = false;
+    }}
+  />
 </section>
